+++ src/test/java/columnar/ColumnarFileTest.java
@@ -20,7 +20,7 @@ public class ColumnarFileTest {
         for (int i = 0; i < 20; i++) {
             attrTypes[i] = new AttrType();
             attrTypes[i].setColumnId(i);
-            attrTypes[i].setSize(4);
+            attrTypes[i].setSize((short)4);
             attrTypes[i].setAttrType(0);
             attrTypes[i].setAttrName("Column" + i);
             in[i] = (int) (Math.random() * 40);
